Baserow structuring and listing rows
