+++ components/baserow/app/baserow.ts
@@ -1,73 +1,42 @@
 import { defineApp } from "@pipedream/types";
 import { axios } from "@pipedream/platform";
-import {
-  AddRecordParams,
-  DataStoreField,
-  HttpRequestParams,
-  UpdateTimerTargetDateParams,
-} from "../common/types";
+import { HttpRequestParams } from "../common/types";
 
 export default defineApp({
   type: "app",
-  app: "niftyimages",
+  app: "baserow",
   methods: {
-    getFieldLabel({
-      name, unique,
-    }: DataStoreField) {
-      return name + (unique
-        ? " (unique)"
-        : "");
-    },
-    getFieldPropType(type: string) {
-      switch (type) {
-      case "NUMBER":
-        return "integer";
-      case "BOOLEAN":
-        return "boolean";
-      default:
-        return "string";
-      }
-    },
     _apiKey(): string {
-      return this.$auth.api_key;
+      return this.$auth.token;
     },
     _baseUrl() {
-      return "https://api.niftyimages.com/v1";
+      return "https://api.baserow.io/api";
     },
     async _httpRequest({
       $ = this,
-      endpoint,
       apiKey,
       ...args
     }: HttpRequestParams): Promise<object> {
       return axios($, {
-        url: this._baseUrl() + endpoint,
+        baseURL: this.baseUrl(),
         headers: {
-          "ApiKey": apiKey || this._apiKey(),
-          "Content-Type": "application/json",
+          "Authorization": `Token ${this._apiKey()}`,
         },
         ...args,
       });
     },
-    async getDataStoreFields(apiKey: string): Promise<DataStoreField[]> {
+    async listRows({ tableId, ...params }): Promise<object> {
       return this._httpRequest({
-        endpoint: "/Store",
-        apiKey,
-      });
-    },
-    async addRecord(args: AddRecordParams): Promise<object> {
-      return this._httpRequest({
-        endpoint: "/Store/AddRecord",
-        method: "POST",
-        ...args,
-      });
-    },
-    async updateTimerTargetDate(args: UpdateTimerTargetDateParams): Promise<object> {
-      return this._httpRequest({
-        endpoint: "/Timer/Update",
-        method: "PUT",
-        ...args,
+        endpoint: `/database/rows/table/${tableId}/`,
+        ...params
       });
     },
   },
+  propDefinitions: {
+    tableId: {
+      label: "Table ID",
+      description: "The ID of the table to use. You can find your tables and their IDs on the [Baserow API Docs](https://baserow.io/api-docs).",
+      type: "integer"
+    }
+  }
 });
