--- components/baserow/baserow.app.mjs
@@ -1,11 +1,0 @@
-export default {
-  type: "app",
-  app: "baserow",
-  propDefinitions: {},
-  methods: {
-    // this.$auth contains connected account data
-    authKeys() {
-      console.log(Object.keys(this.$auth));
-    },
-  },
-};
