+++ components/baserow/common/types.ts
@@ -7,29 +7,7 @@ export interface HttpRequestParams {
   apiKey?: string;
 }
 
-export interface DataStoreField {
-  name: string;
-  type: string;
-  date_input_format?: string;
-  unique: boolean;
-}
-
-export interface AddRecordParams {
+export interface ListRowsParams {
   $: Pipedream;
-  apiKey: string;
-  data: object;
-}
-
-export interface UpdateTimerTargetDateParams {
-  $: Pipedream;
-  apiKey: string;
-  data: {
-    TimerImageUrl: string;
-    TargetDate: string;
-    Format?: string;
-    IsUtc?: boolean;
-    AddHours?: number;
-    AddDays?: number;
-    AddMonths?: number;
-  };
+  tableId: number;
 }
